+++ examples/src/section.rs
